requirements.txt
# dashboard requirements
# Usage: pip install -r requirements.txt

bottle==0.12.25
faker
# paho-mqtt==1.5.1
paho-mqtt
#jsons
python-dotenv==1.0.0
mysql-connector-python==8.2.0
jinja2==3.1.2








